+++ db/pvault.sql
@@ -8,6 +8,7 @@ CREATE TABLE `ingredient_compounds` (
   `ec` varchar(255) COLLATE utf8_general_ci DEFAULT NULL,
   `percentage` DECIMAL(8,4) NOT NULL,
   `toDeclare` INT NOT NULL DEFAULT '0',
+  `GHS` TEXT NOT NULL DEFAULT '-',
   `created` timestamp NOT NULL DEFAULT current_timestamp() ON UPDATE current_timestamp()
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;