+++ db/updates/update_10.1-10.2.sql
@@ -1,1 +1,2 @@
 RENAME TABLE `allergens` TO `ingredient_compounds`;
+ALTER TABLE `ingredient_compounds` ADD `GHS` TEXT NOT NULL DEFAULT '-' AFTER `toDeclare`;