+++ pages/views/ingredients/compos.php
@@ -15,10 +15,10 @@ $ingName = mysqli_real_escape_string($conn, $_GET["name"]);
 <div class="card-body">
  	<div class="text-right">
   		<div class="btn-group">
-   			<button type="button" class="btn btn-primary dropdown-toggle" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><i class="fa fa-bars mr2"></i>Actions</button>
+   			<button type="button" class="btn btn-primary dropdown-toggle" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><i class="fa fa-bars mx-2"></i>Actions</button>
     		<div class="dropdown-menu dropdown-menu-right">
-        		<li><a class="dropdown-item" href="#" data-bs-toggle="modal" data-bs-target="#addComposition"><i class="fa-solid fa-plus mr2"></i>Add new</a></li>
-        		<li><a class="dropdown-item" href="#" data-bs-toggle="modal" data-bs-target="#addCSV"><i class="fa-solid fa-file-import mr2"></i>Upload CSV</a></li>
+        		<li><a class="dropdown-item" href="#" data-bs-toggle="modal" data-bs-target="#addComposition"><i class="fa-solid fa-plus mx-2"></i>Add new</a></li>
+        		<li><a class="dropdown-item" href="#" data-bs-toggle="modal" data-bs-target="#addCSV"><i class="fa-solid fa-file-import mx-2"></i>Upload CSV</a></li>
     		</div>
   		</div>                    
 	</div>
@@ -31,6 +31,7 @@ $ingName = mysqli_real_escape_string($conn, $_GET["name"]);
           <th>CAS</th>
           <th>EINECS</th>
           <th>Percentage</th>
+          <th>GHS Classification</th>
           <th>Declare</th>
           <th></th>
       </tr>
@@ -45,7 +46,7 @@ $('[data-bs-toggle="tooltip"]').tooltip();
 var tdCompositions = $('#tdCompositions').DataTable( {
 	columnDefs: [
 		{ className: 'text-center', targets: '_all' },
-		{ orderable: false, targets: [5] },
+		{ orderable: false, targets: [4,6] },
 	],
 	dom: 'lfrtip',
 	processing: true,
@@ -61,6 +62,7 @@ var tdCompositions = $('#tdCompositions').DataTable( {
 			  { data : 'cas', title: 'CAS', render: cmpCAS},
 			  { data : 'ec', title: 'EINECS', render: cmpEC},
 			  { data : 'percentage', title: 'Percentage', render: cmpPerc},
+			  { data : 'GHS', title: 'GHS Classification', render: cmpGHS},
 			  { data : 'toDeclare', title: 'Declare', render: cmpDeclare},
 			  { data : null, title: '', render: cmpActions},		   
 			 ],
@@ -85,6 +87,10 @@ function cmpEC(data, type, row){
 
 function cmpPerc(data, type, row){
 	return '<i class="percentage pv_point_gen" data-name="percentage" data-type="text" data-pk="'+row.id+'">'+row.percentage+'</i>';    
+}
+
+function cmpGHS(data, type, row){
+	return '<i class="GHS pv_point_gen" data-name="GHS" data-type="text" data-pk="'+row.id+'">'+row.GHS+'</i>';    
 }
 
 function cmpDeclare(data, type, row){
@@ -132,6 +138,13 @@ $('#tdCompositions').editable({
     title: 'Percentage'
 });
 
+$('#tdCompositions').editable({
+    container: 'body',
+    selector: 'i.GHS',
+    type: 'POST',
+    url: "update_data.php?composition=update&ing=<?=$ingName;?>",
+    title: 'GHS'
+});
 
 $('#tdCompositions').editable({
 	pvnoresp: false,
@@ -205,6 +218,7 @@ $('#addComposition').on('click', '[id*=cmpAdd]', function () {
 			allgPerc: $("#allgPerc").val(),
 			allgCAS: $("#allgCAS").val(),
 			allgEC: $("#allgEC").val(),	
+			GHS: $("#GHS").val(),	
 			addToIng: $("#addToIng").is(':checked'),
 			addToDeclare: $("#addToDeclare").is(':checked'),
 			ing: '<?=$ingName?>'
@@ -217,6 +231,7 @@ $('#addComposition').on('click', '[id*=cmpAdd]', function () {
 				$("#allgCAS").val('');
 				$("#allgEC").val('');
 				$("#allgPerc").val('');
+				$("#GHS").val('');
 				reload_cmp_data();
 			}else{
 				var msg = '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-bs-dismiss="alert" aria-label="close">x</a>' + data.error + '</div>';
@@ -276,32 +291,38 @@ function reload_cmp_data() {
         </button>
       </div>
       <div class="modal-body">
-      <div id="inf"></div>
-            Name: 
-            <input class="form-control" name="allgName" type="text" id="allgName" />
-            <p>
-            CAS: 
-            <input class="form-control" name="allgCAS" type="text" id="allgCAS" />
-            <p>
-            EINECS: 
-            <input class="form-control" name="allgEC" type="text" id="allgEC" />
-            <p>            
-            Percentage %:
-            <input class="form-control" name="allgPerc" type="text" id="allgPerc" />
-            </p>
-            <div class="dropdown-divider"></div>
-          <p>
-          <label>
-             <input name="addToDeclare" type="checkbox" id="addToDeclare" value="1" />
-            To declare in warnings
-          </label>
-          </p>
-          <p>
-          <label>
-             <input name="addToIng" type="checkbox" id="addToIng" value="1" />
-            Add to ingredients
-          </label>
-      </div>
+        <div id="inf"></div>
+        <div class="mb-3">
+	        <label for="allgName" class="form-label">Name</label>
+    	    <input class="form-control" name="allgName" type="text" id="allgName" />
+        </div>
+        <div class="mb-3">
+        	<label for="allgCAS" class="form-label">CAS</label>
+        	<input class="form-control" name="allgCAS" type="text" id="allgCAS" />
+        </div>
+        <div class="mb-3">
+        	<label for="allgEC" class="form-label">EINECS</label>
+        	<input class="form-control" name="allgEC" type="text" id="allgEC" />
+        </div>
+        <div class="mb-3">
+	        <label for="allgPerc" class="form-label">Percentage</label>
+    	    <input class="form-control" name="allgPerc" type="text" id="allgPerc" />
+        </div>
+        <div class="mb-3">
+	        <label for="GHS" class="form-label">GHS Classification</label>
+    	    <input class="form-control" name="GHS" type="text" id="GHS" />
+        </div>        
+        <hr class="dropdown-divider" />
+        <div class="form-check">
+            <input class="form-check-input" name="addToDeclare" type="checkbox" id="addToDeclare" value="1" />
+            <label class="form-check-label" for="addToDeclare">To declare in warnings</label>
+        </div>
+        <div class="form-check">
+            <input class="form-check-input" name="addToIng" type="checkbox" id="addToIng" value="1" />
+            <label class="form-check-label" for="addToIng">Add to ingredients</label>
+        </div>
+    </div>
+
       <div class="modal-footer">
         <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
         <input type="submit" name="button" class="btn btn-primary" id="cmpAdd" value="Add">
@@ -322,25 +343,17 @@ function reload_cmp_data() {
         </button>
       </div>
       <div class="modal-body">
+      
       <div id="CSVImportMsg"></div>
-        <table width="100%" border="0">
-          <tr>
-            <td width="22%">Choose file:</td>
-            <td width="78%">
-              <input type="file" name="CSVFile" id="CSVFile" class="form-control" />
-            </td>
-          </tr>
-          <tr>
-            <td colspan="2"><hr />
-              <p>CSV format: <strong>ingredient,CAS,EINECS,percentage</strong></p>
-            <p>Example: <em><strong>Citral,5392-40-5,226-394-6,0.15</strong></em></p>
-            <p>Duplicates will be ignored.</p></td>
-          </tr>
-          <tr>
-            <td>&nbsp;</td>
-            <td>&nbsp;</td>
-          </tr>
-        </table>
+     	<div class="mb-3">
+  			<label for="formFile" class="form-label">Choose file</label>
+            <input type="file" name="CSVFile" id="CSVFile" class="form-control" />
+        </div>
+        <hr class="dropdown-divider" />        
+        <p>CSV format: <strong>ingredient,CAS,EINECS,percentage,GHS</strong></p>
+        <p>Example: <em><strong>Citral,5392-40-5,226-394-6,0.15,Skin Irrit. 2-Eye Irrit</strong></em></p>
+        <p>Duplicates will be ignored.</p>
+            
 	  <div class="modal-footer">
         <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
         <input type="submit" name="cmpCSV" class="btn btn-primary" id="cmpCSV" value="Import">