+++ releasenotes.md
@@ -1,4 +1,5 @@
 Whats New in v10.2
 --------------------------
+- Added GHS Classification for ingredient compounds
 - Added JSON export for formula making
 - For full details please refer to the CHANGELOG